+++ src/App.jsx
@@ -1,14 +1,14 @@
 import React, { useState, useRef, useEffect } from 'react';
 import { BrowserRouter as Router, Routes, Route} from "react-router-dom"
-import SearchBar from './components/SearchBar.jsx';
-import PlantList from './components/PlantList.jsx';
-import Footer from './components/Footer.jsx';
-import LoginPage from './components/LoginPage.jsx';
-import PlantDashboard from './components/PlantDashBoard.jsx';
-import PrivateRoute from './components/PrivateRoute.jsx';
-import { AuthProvider } from './components/AuthContext';
+import SearchBar from './components/SearchBar/SearchBar.jsx';
+import PlantList from './components/PlantList/PlantList.jsx';
+import Footer from './components/Footer/Footer.jsx';
+import LoginPage from './components/Authentication/LoginPage/LoginPage.jsx';
+import PlantDashboard from './components/PlantDashboard/PlantDashboard.jsx';
+import PrivateRoute from './components/Authentication/PrivateRoute.jsx';
+import { AuthProvider } from './components/Authentication/AuthContext.jsx';
 import { ToastContainer } from 'react-toastify';
-import Header from './components/Header.jsx';
+import Header from './components/Header/Header.jsx';
 import styles from './App.module.css';
 
 const App = () => {
