+++ src/components/Authentication/AuthContext.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, useContext } from 'react';
-import { auth } from '../firebase/firebase.js';
+import { auth } from '../../firebase/firebase.js';
 import { onAuthStateChanged } from 'firebase/auth';
 
 const AuthContext = React.createContext();
