+++ src/components/Authentication/LoginLogoutButton/LoginLogoutButton.jsx
@@ -1,6 +1,6 @@
 import { useNavigate } from "react-router-dom";
-import { doSignOut } from "../firebase/auth.js";
-import { useAuth } from "./AuthContext.jsx";
+import { doSignOut } from "../../../firebase/auth.js";
+import { useAuth } from "../AuthContext.jsx";
 import styles from './LoginLogoutButton.module.css'
 
 const LoginLogoutButton = () => {
