+++ src/components/PlantDetails/PlantDetails.jsx
