+++ src/components/PlantDetails/PlantDetails.module.css
