+++ src/components/PlantDetails/PlantDetailsData.jsx
